update bug free modules for tracklet connectivity

+++ datasets/TrackletpairDataset.py
@@ -181,10 +181,8 @@ class TrackletpairDataset(Dataset):
 
         tracklet_mask_1 = torch.from_numpy(np.array(tracklet_mask_1).astype(np.float32))
         tracklet_mask_2 = torch.from_numpy(np.array(tracklet_mask_2).astype(np.float32))
-        connectivity = torch.from_numpy(np.array(int(connectivity))).reshape(-1, 1)
+        connectivity = torch.from_numpy(np.array(int(connectivity)).astype(np.float32)).reshape(-1, 1)
         
-        
-
         return img_1, img_2, loc_mat, tracklet_mask_1, tracklet_mask_2, real_window_len, connectivity
